requirements.txt
# Core dependencies
opencv-python>=4.8.1
numpy>=1.26
PyYAML>=6.0.1

# YOLOv8 (also pulls torch, torchvision, Pillow, matplotlib, tqdm, etc.)
ultralytics==8.4.21

# PyTorch (explicit pin to ensure CUDA build; ultralytics requires it)
torch>=2.1.0
